Implement the challenge requirements
If the payment is for a physical item, you must generate one shipping
labelfor it to be placed in the shipping box;

If payment is a service subscription, you must activate the
subscription, and notify the user via email about this;

If the payment is an ordinary book, you must generate it shipping label
with a notification that it is a tax-exempt item as provided in the
Constitution Art. 150, VI, d.

If payment of any digital media (music, video), in addition to sending
the description of the purchase by email to the buyer, grant a discount
voucher of $ 10 to the buyer associated with the payment.