Extract Payment class from bootstrap to a new file
Remove Payment class from 'bootrstrap.rb' and create a new file for it
on the project tree (lib/models). The @payment_method assignment on the
'initialize' method has been changed to avoid assigning an array to the
instance variable (Hash#values_at always returns an array, even if a
single argument is provided).

Add file requirements loading to bootstrap.

Create test/spec file for Payment class as well as a spec helper for
loading the project dependencies.